Tpetra: doPosts Converted to Use Kokkos
All doPost related functions in DistributorActor now use Kokkos Views instead of Teuchos ArrayViews.
getImportViewLimits and getExportViewLimits now use parallel loops.
All calls to these functions have been modified accordingly.
Fence count has increased for unknown reasons. See PR for more details.

+++ packages/tpetra/core/test/Distributor/Distributor_UnitTests.cpp
@@ -740,8 +740,8 @@ namespace {
     Tpetra::Details::DistributorActor actor;
 
     // Perform communication steps
-    actor.doPostRecvs(plan, 1, imports);
-    actor.doPostSends(plan, myExportsConst, 1, imports);
+    actor.doPostRecvs(plan, (size_t) 1, imports);
+    actor.doPostSends(plan, myExportsConst, (size_t) 1, imports);
     actor.doWaitsRecv(plan);
     actor.doWaitsSend(plan);
 
@@ -858,8 +858,8 @@ namespace {
       Tpetra::Details::DistributorActor actor;
 
       // Perform communication steps
-      actor.doPostRecvs(plan, 1, imports);
-      actor.doPostSends(plan, myExportsConst, 1, imports);
+      actor.doPostRecvs(plan, (size_t) 1, imports);
+      actor.doPostSends(plan, myExportsConst, (size_t) 1, imports);
       actor.doWaitsRecv(plan);
       actor.doWaitsSend(plan);
